Optoro hackday work (#2)
Here's what we did/what is in each directory:

Vacant buildings

Extracted data from PDFs using tabula, and threw it on a google map. Jupyter notebook in etl directory in that folder.
DC Main Streets

Merge the smaller excel files into a way that made sense/recorded our questions
Combined the directory of many CSVs into one big CSV, with a column on which file they came from. Also included the script that makes this happen.
Department of Corrections

There are two threads here:
The main repo has a readme with some questions about the internal consistency of the data
The weekly_corrections_data folder has the weekly corrections data in various forms: csv, excel, and the original PDFs
Rat hacking

Playing around with the 311 data about rats
Mayor's Signed

We couldn't figure out what to do with this and it just seemed like a couple of letters